unnumbered clauses: https://github.com/metanorma/metanorma-standoc/issues/887
