+++ spec/isodoc/section_spec.rb
@@ -124,34 +124,6 @@ RSpec.describe IsoDoc do
       OUTPUT
   end
 
-  it "processes document with no content" do
-    input = <<~INPUT
-      <iso-standard xmlns="http://riboseinc.com/isoxml">
-            <preface/>
-            <sections/>
-          </iso-standard>
-    INPUT
-    output = <<~OUTPUT
-          <html xmlns:epub="http://www.idpf.org/2007/ops" lang="en">
-        <head/>
-        <body lang="en">
-          <div class="title-section">
-            <p>&#160;</p>
-          </div>
-          <br/>
-          <div class="prefatory-section">
-            <p>&#160;</p>
-          </div>
-          <br/>
-          <div class="main-section">
-          </div>
-        </body>
-      </html>
-    OUTPUT
-    expect(xmlpp(IsoDoc::HtmlConvert.new({})
-.convert("test", input, true))).to be_equivalent_to xmlpp(output)
-  end
-
   it "processes section names" do
     input = <<~INPUT
       <iso-standard xmlns="http://riboseinc.com/isoxml">
@@ -570,22 +542,18 @@ RSpec.describe IsoDoc do
             <div class="authority">
           <div class="boilerplate-copyright">
             <div><h1>Copyright</h1>
-
             </div>
           </div>
           <div class="boilerplate-license">
             <div><h1>License</h1>
-
             </div>
           </div>
           <div class="boilerplate-legal">
             <div><h1>Legal</h1>
-
             </div>
           </div>
           <div class="boilerplate-feedback">
             <div><h1>Feedback</h1>
-
             </div>
           </div>
         </div>
@@ -614,7 +582,6 @@ RSpec.describe IsoDoc do
             <div class="Section3" id="B">
               <h1 class="IntroTitle">Introduction</h1>
               <div id="C"><h2>Introduction Subsection</h2>
-
             </div>
             </div>
             <p class="page-break">
@@ -652,19 +619,14 @@ RSpec.describe IsoDoc do
               <p id="E">Text</p>
             </div>
             <div><h1>2.<span style="mso-tab-count:1">&#160; </span>Normative References</h1>
-
           </div>
             <div id="H">
               <h1>3.<span style="mso-tab-count:1">&#160; </span>Terms, Definitions, Symbols and Abbreviated Terms</h1>
               <div id="I"><h2>3.1.<span style="mso-tab-count:1">&#160; </span>Normal Terms</h2>
-
               <p class="TermNum" id="J">3.1.1.</p>
-
                 <p class="Terms" style="text-align:left;"><b>Term2</b></p>
-
             </div>
               <div id="K"><h2>3.2.<span style="mso-tab-count:1">&#160; </span>Definitions</h2>
-
               <table class="dl"><tr><td valign="top" align="left"><p align="left" style="margin-left:0pt;text-align:left;">Symbol</p></td><td valign="top">Definition</td></tr></table>
             </div>
             </div>
@@ -682,13 +644,10 @@ RSpec.describe IsoDoc do
             <div id="M">
               <h1>5.<span style="mso-tab-count:1">&#160; </span>Clause 4</h1>
               <div id="N"><h2>5.1.<span style="mso-tab-count:1">&#160; </span>Introduction</h2>
-
             </div>
               <div id="O"><h2>5.2.<span style="mso-tab-count:1">&#160; </span>Clause 4.2</h2>
-
             </div>
               <div id="O1"><h2>5.3.</h2>
-
             </div>
             </div>
             <p class="page-break">
@@ -698,12 +657,9 @@ RSpec.describe IsoDoc do
               <h1 class="Annex"> <b>Annex A</b><br/>(normative)<br/><br/><b>Annex</b>
           </h1>
               <div id="Q"><h2>A.1.<span style="mso-tab-count:1">&#160; </span>Annex A.1</h2>
-
             <div id="Q1"><h3>A.1.1.<span style="mso-tab-count:1">&#160; </span>Annex A.1a</h3>
-
             </div>
             <div><h3 class="Section3">A.1.2.<span style="mso-tab-count:1">&#160; </span>Annex Bibliography</h3>
-
             </div>
           </div>
             </div>
@@ -717,9 +673,7 @@ RSpec.describe IsoDoc do
               <br clear="all" style="mso-special-character:line-break;page-break-before:always"/>
             </p>
             <div><h1 class="Section3">Bibliography</h1>
-
             <div><h2 class="Section3">Bibliography Subsection</h2>
-
             </div>
           </div>
                 <p class="page-break">
@@ -1549,6 +1503,147 @@ RSpec.describe IsoDoc do
       .convert("test", input, true)))).to be_equivalent_to xmlpp(output)
   end
 
+  it "processes unnumbered section names" do
+    input = <<~INPUT
+      <iso-standard xmlns="http://riboseinc.com/isoxml">
+       <sections>
+       <clause id="D" obligation="normative" type="scope" unnumbered="true">
+         <title>Scope</title>
+         <p id="E">Text</p>
+       </clause>
+       <clause id="H" obligation="normative" unnumbered="true"><title>Terms, Definitions, Symbols and Abbreviated Terms</title><terms id="I" obligation="normative">
+         <title>Normal Terms</title>
+         <term id="J">
+         <preferred><expression><name>Term2</name></expression></preferred>
+       </term>
+       </terms>
+       <definitions id="K" unnumbered="true">
+         <dl>
+         <dt>Symbol</dt>
+         <dd>Definition</dd>
+         </dl>
+       </definitions>
+       </clause>
+       <definitions id="L" unnumbered="true">
+         <dl>
+         <dt>Symbol</dt>
+         <dd>Definition</dd>
+         </dl>
+       </definitions>
+       <clause id="M" inline-header="false" obligation="normative" unnumbered="true"><title>Clause 4</title>
+        <clause id="N" inline-header="false" obligation="normative">
+         <title>Introduction</title>
+        </clause>
+       <clause id="O" inline-header="false" obligation="normative">
+         <title>Clause 4.2</title>
+       </clause>
+        </clause>
+       <clause id="O1" inline-header="false" obligation="normative">
+       </clause>
+       <clause id="O2" inline-header="false" obligation="normative" unnumbered="true">
+       </clause>
+       <clause id="O3" inline-header="false" obligation="normative">
+       </clause>
+
+       </sections><annex id="P" inline-header="false" obligation="normative" unnumbered="true">
+         <title>Annex</title>
+         <clause id="Q" inline-header="false" obligation="normative">
+         <title>Annex A.1</title>
+         <clause id="Q1" inline-header="false" obligation="normative">
+         <title>Annex A.1a</title>
+         </clause>
+       </clause>
+       </annex><bibliography><references id="R" obligation="informative" normative="true" unnumbered="true">
+         <title>Normative References</title>
+       </references><clause id="S" obligation="informative" unnumbered="true">
+         <title>Bibliography</title>
+         <references id="T" obligation="informative" normative="false">
+         <title>Bibliography Subsection</title>
+       </references>
+       </clause>
+       </bibliography>
+       </iso-standard>
+    INPUT
+    output = <<~OUTPUT
+       <iso-standard xmlns="http://riboseinc.com/isoxml" type="presentation">
+         <preface>
+           <clause type="toc" id="_" displayorder="1">
+             <title depth="1">Table of contents</title>
+           </clause>
+         </preface>
+         <sections>
+           <clause id="D" obligation="normative" type="scope" unnumbered="true" displayorder="2">
+             <title depth="1">Scope</title>
+             <p id="E">Text</p>
+           </clause>
+           <clause id="H" obligation="normative" unnumbered="true" displayorder="4">
+             <title depth="1">Terms, Definitions, Symbols and Abbreviated Terms</title>
+             <terms id="I" obligation="normative">
+               <title depth="2">Normal Terms</title>
+               <term id="J">
+                 <preferred>
+                   <strong>Term2</strong>
+                 </preferred>
+               </term>
+             </terms>
+             <definitions id="K" unnumbered="true">
+               <dl>
+                 <dt>Symbol</dt>
+                 <dd>Definition</dd>
+               </dl>
+             </definitions>
+           </clause>
+           <definitions id="L" unnumbered="true" displayorder="5">
+             <dl>
+               <dt>Symbol</dt>
+               <dd>Definition</dd>
+             </dl>
+           </definitions>
+           <clause id="M" inline-header="false" obligation="normative" unnumbered="true" displayorder="6">
+             <title depth="1">Clause 4</title>
+             <clause id="N" inline-header="false" obligation="normative">
+               <title depth="2">Introduction</title>
+             </clause>
+             <clause id="O" inline-header="false" obligation="normative">
+               <title depth="2">Clause 4.2</title>
+             </clause>
+           </clause>
+           <clause id="O1" inline-header="false" obligation="normative" displayorder="7">
+             <title>1.</title>
+           </clause>
+           <clause id="O2" inline-header="false" obligation="normative" unnumbered="true" displayorder="8">
+        </clause>
+           <clause id="O3" inline-header="false" obligation="normative" displayorder="9">
+             <title>2.</title>
+           </clause>
+           <references id="R" obligation="informative" normative="true" unnumbered="true" displayorder="3">
+             <title depth="1">Normative References</title>
+           </references>
+         </sections>
+         <annex id="P" inline-header="false" obligation="normative" unnumbered="true" displayorder="10">
+           <title><strong>Annex </strong><br/>(normative)<br/><br/><strong>Annex</strong></title>
+           <clause id="Q" inline-header="false" obligation="normative">
+             <title depth="2">.1.<tab/>Annex A.1</title>
+             <clause id="Q1" inline-header="false" obligation="normative">
+               <title depth="3">.1.1.<tab/>Annex A.1a</title>
+             </clause>
+           </clause>
+         </annex>
+         <bibliography>
+           <clause id="S" obligation="informative" unnumbered="true" displayorder="11">
+             <title depth="1">Bibliography</title>
+             <references id="T" obligation="informative" normative="false">
+               <title depth="2">Bibliography Subsection</title>
+             </references>
+           </clause>
+         </bibliography>
+       </iso-standard>
+    OUTPUT
+    expect(xmlpp(strip_guid(IsoDoc::PresentationXMLConvert
+      .new(presxml_options)
+      .convert("test", input, true)))).to be_equivalent_to xmlpp(output)
+  end
+
   it "processes floating titles" do
     input = <<~INPUT
           <iso-standard xmlns="http://riboseinc.com/isoxml">
@@ -1747,54 +1842,6 @@ RSpec.describe IsoDoc do
       .convert("test", input, true)))).to be_equivalent_to xmlpp(output)
   end
 
-  it "processes simple terms & definitions" do
-    input = <<~INPUT
-              <iso-standard xmlns="http://riboseinc.com/isoxml">
-      <sections>
-      <terms id="H" obligation="normative"><title>Terms, Definitions, Symbols and Abbreviated Terms</title>
-        <term id="J">
-        <preferred><expression><name>Term2</name></expression></preferred>
-      </term>
-       </terms>
-       </sections>
-       </iso-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-      <iso-standard xmlns='http://riboseinc.com/isoxml' type="presentation">
-        <preface>
-          <clause type="toc" id="_" displayorder="1">
-           <title depth="1">Table of contents</title>
-          </clause>
-        </preface>
-        <sections>
-          <terms id='H' obligation='normative' displayorder="2">
-          <title depth='1'>1.<tab/>Terms, Definitions, Symbols and Abbreviated Terms</title>
-            <term id='J'>
-            <name>1.1.</name>
-              <preferred><strong>Term2</strong></preferred>
-            </term>
-          </terms>
-        </sections>
-      </iso-standard>
-    OUTPUT
-
-    html = <<~"OUTPUT"
-      #{HTML_HDR}
-                   <div id="H"><h1>1.&#160; Terms, Definitions, Symbols and Abbreviated Terms</h1>
-           <p class="TermNum" id="J">1.1.</p>
-             <p class="Terms" style="text-align:left;"><b>Term2</b></p>
-           </div>
-                 </div>
-               </body>
-           </html>
-    OUTPUT
-    expect(xmlpp(strip_guid(IsoDoc::PresentationXMLConvert.new(presxml_options)
-      .convert("test", input, true)))).to be_equivalent_to xmlpp(presxml)
-    expect(xmlpp(IsoDoc::HtmlConvert.new({})
-      .convert("test", presxml, true))).to be_equivalent_to xmlpp(html)
-  end
-
   it "processes inline section headers" do
     input = <<~INPUT
       <iso-standard xmlns="http://riboseinc.com/isoxml">
@@ -1937,94 +1984,6 @@ RSpec.describe IsoDoc do
     OUTPUT
     expect(xmlpp(strip_guid(IsoDoc::PresentationXMLConvert.new(presxml_options)
       .convert("test", input, true)))).to be_equivalent_to xmlpp(output)
-  end
-
-  it "processes clauses containing normative references" do
-    input = <<~INPUT
-           <iso-standard xmlns="http://riboseinc.com/isoxml">
-           <bibliography>
-       <clause id="D" obligation="informative">
-        <title>Bibliography</title>
-        <references id="E" obligation="informative" normative="false">
-        <title>Bibliography Subsection 1</title>
-      </references>
-        <references id="F" obligation="informative" normative="false">
-        <title>Bibliography Subsection 2</title>
-      </references>
-      </clause>
-      <clause id="A" obligation="informative"><title>First References</title>
-       <references id="B" obligation="informative" normative="true">
-        <title>Normative References 1</title>
-      </references>
-       <references id="C" obligation="informative" normative="false">
-        <title>Normative References 2</title>
-      </references>
-       </clause>
-
-      </bibliography>
-      </iso-standard>
-    INPUT
-    presxml = <<~OUTPUT
-      <iso-standard xmlns="http://riboseinc.com/isoxml" type="presentation">
-         <preface>
-           <clause type="toc" id="_" displayorder="1">
-             <title depth="1">Table of contents</title>
-           </clause>
-         </preface>
-         <sections>
-           <clause id="A" obligation="informative" displayorder="2">
-             <title depth="1">1.<tab/>First References</title>
-             <references id="B" obligation="informative" normative="true">
-               <title depth="2">1.1.<tab/>Normative References 1</title>
-             </references>
-             <references id="C" obligation="informative" normative="false">
-               <title depth="2">1.2.<tab/>Normative References 2</title>
-             </references>
-           </clause>
-         </sections>
-         <bibliography>
-           <clause id="D" obligation="informative" displayorder="3">
-             <title depth="1">Bibliography</title>
-             <references id="E" obligation="informative" normative="false">
-               <title depth="2">Bibliography Subsection 1</title>
-             </references>
-             <references id="F" obligation="informative" normative="false">
-               <title depth="2">Bibliography Subsection 2</title>
-             </references>
-           </clause>
-         </bibliography>
-       </iso-standard>
-    OUTPUT
-
-    html = <<~OUTPUT
-      #{HTML_HDR}
-                 <div>
-                   <h1>1.&#160; First References</h1>
-                   <div>
-                     <h2 class='Section3'>1.1.&#160; Normative References 1</h2>
-                   </div>
-                   <div>
-                     <h2 class='Section3'>1.2.&#160; Normative References 2</h2>
-                   </div>
-                 </div>
-                 <br/>
-                 <div>
-                   <h1 class='Section3'>Bibliography</h1>
-                   <div>
-                     <h2 class='Section3'>Bibliography Subsection 1</h2>
-                   </div>
-                   <div>
-                     <h2 class='Section3'>Bibliography Subsection 2</h2>
-                   </div>
-                 </div>
-               </div>
-             </body>
-           </html>
-    OUTPUT
-    expect(xmlpp(strip_guid(IsoDoc::PresentationXMLConvert.new(presxml_options)
-      .convert("test", input, true)))).to be_equivalent_to xmlpp(presxml)
-    expect(xmlpp(IsoDoc::HtmlConvert.new({})
-      .convert("test", presxml, true))).to be_equivalent_to xmlpp(html)
   end
 
   it "processes annexes containing one, or more than one special sections" do
